### 2022-07-01 #4535 @editorialbot
[PRE REVIEW]: Mallob: Scalable SAT Solving On Demand With Decentralized Job Scheduling
**Submitting author:** @domschrei (<a href="http://orcid.org/0000-0002-4185-1851">Dominik Schreiber</a>)
**Repository:** https://github.com/domschrei/mallob
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.0
