+++ src/app/kmeans/kmeans_utils.hpp
@@ -4,6 +4,6 @@
 
 namespace KMeansUtils {
     typedef std::vector<float> Point;
-    float eukild(Point& p1, Point& p2);
+    float eukild(const float* p1, Point& p2);
     std::vector<int> childIndexesOf(int parentIndex, int jobVolume);
 };  // namespace KMeansUtils
