### 2022-07-05 1232409
Fix naming of registry method
