### 2022-07-05 #4535 @danielskatz
👋 @ARMartinelli - I think you are willing to review this JOSS submission based on a EuroPar paper, correct?
