### 2022-07-05 2d7c228
no kMeansData anymore
