+++ paper/paper.bib
@@ -202,7 +202,8 @@
   booktitle={International Conference on Theory and Applications of Satisfiability Testing},
   pages={228--245},
   year={2016},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-319-40970-2_15}
 }
 
 @inproceedings{huang2003adaptive,
