### 2022-07-06 9ea9767
Towards improved termination & interruption protocol
