+++ src/worker.cpp
@@ -72,16 +72,14 @@ void Worker::init() {
     // Begin listening to incoming messages
     q.registerCallback(MSG_ANSWER_ADOPTION_OFFER,
         [&](auto& h) {handleAnswerAdoptionOffer(h);});
-    q.registerCallback(MSG_NOTIFY_JOB_ABORTING, 
-        [&](auto& h) {handleNotifyJobAborting(h);});
-    q.registerCallback(MSG_NOTIFY_JOB_TERMINATING, 
-        [&](auto& h) {handleNotifyJobTerminating(h);});
+    q.registerCallback(MSG_CANCEL_JOB, 
+        [&](auto& h) {handleCancelJob(h);});
     q.registerCallback(MSG_NOTIFY_RESULT_FOUND, 
         [&](auto& h) {handleNotifyResultFound(h);});
     q.registerCallback(MSG_INCREMENTAL_JOB_FINISHED,
         [&](auto& h) {handleIncrementalJobFinished(h);});
-    q.registerCallback(MSG_INTERRUPT,
-        [&](auto& h) {handleInterrupt(h);});
+    q.registerCallback(MSG_CLOSE_JOB_REVISION,
+        [&](auto& h) {handleCloseJobRevision(h);});
     q.registerCallback(MSG_NOTIFY_NODE_LEAVING_JOB, 
         [&](auto& h) {handleNotifyNodeLeavingJob(h);});
     q.registerCallback(MSG_NOTIFY_RESULT_OBSOLETE, 
@@ -430,12 +428,12 @@ void Worker::publishAndResetSysState() {
     _sys_state.setLocal(SYSSTATE_SUMDESIRELATENCIES, 0);
 }
 
-void Worker::handleNotifyJobAborting(MessageHandle& handle) {
+void Worker::handleCancelJob(MessageHandle& handle) {
 
-    int jobId = Serializable::get<int>(handle.getRecvData());
+    auto [jobId, revision] = Serializable::get<IntPair>(handle.getRecvData());
     if (!_job_db.has(jobId)) return;
 
-    LOG(V3_VERB, "Acknowledge #%i aborting\n", jobId);
+    LOG(V3_VERB, "Cancelling #%i rev. %i\n", jobId, revision);
     auto& job = _job_db.get(jobId);    
     if (job.getJobTree().isRoot()) {
         // Forward information on aborted job to client
@@ -443,11 +441,7 @@ void Worker::handleNotifyJobAborting(MessageHandle& handle) {
             MSG_NOTIFY_CLIENT_JOB_ABORTING, handle.moveRecvData());
     }
 
-    if (job.isIncremental()) {
-        interruptJob(jobId, /*terminate=*/false, /*reckless=*/false);
-    } else {
-        interruptJob(jobId, /*terminate=*/true, /*reckless=*/true);
-    }
+    propagateRevisionClosed(jobId, revision, /*successfulRank=*/-1);
 }
 
 void Worker::handleAnswerAdoptionOffer(MessageHandle& handle) {
@@ -702,12 +696,8 @@ void Worker::handleIncrementalJobFinished(MessageHandle& handle) {
     int jobId = Serializable::get<int>(handle.getRecvData());
     if (_job_db.has(jobId)) {
         LOG(V3_VERB, "Incremental job %s done\n", _job_db.get(jobId).toStr());
-        interruptJob(Serializable::get<int>(handle.getRecvData()), /*terminate=*/true, /*reckless=*/false);
+        propagateRevisionClosed(jobId, /*revision=*/-1, /*successfulRank=*/-1);
     }
-}
-
-void Worker::handleInterrupt(MessageHandle& handle) {
[…]
