### 2022-07-07 1597fa2
Bugfixing for new incremental solving protocol
