### 2022-07-07 066c85d
Clean up and clarify general execution documentation
