### 2022-07-07 bd5920c
Output proofs in logging directory
