### 2022-07-08 90a7dac
Advance new incremental interface
