### 2022-07-08 c5bf592
Transfer number of original clauses to solvers
