+++ src/app/sat/execution/solver_setup.hpp
@@ -18,6 +18,7 @@ struct SolverSetup {
 	bool hasPseudoincrementalSolvers;
 	char solverType;
 	int solverRevision;
+	int numOriginalClauses;
 	
 	int minNumChunksPerSolver;
 	int numBufferedClsGenerations;
