### 2022-07-08 79aa3b9
Changes to support passing num original clauses to instances (from https://github.com/domschrei/mallob/commit/c5bf592e621b67e6da602b7fa83ff43bac11dd6c)
