+++ src/app/sat/solvers/cadical.cpp
@@ -44,7 +44,7 @@ void Cadical::diversify(int seed) {
 
 	// In certified UNSAT mode?
 	if (getSolverSetup().certifiedUnsat) {
-		
+
 		int solverRank = getSolverSetup().globalId;
 		int maxNumSolvers = getSolverSetup().maxNumSolvers;
 		LOGGER(_logger, V3_VERB, "Diversifying rank=%i size=%i DI=%i with certified UNSAT support\n", 
@@ -52,6 +52,7 @@ void Cadical::diversify(int seed) {
                 //Need to do +1 so we don't start at 0
 		okay = solver->set("instance_num", solverRank + 1); assert(okay);
 		okay = solver->set("total_instances", maxNumSolvers); assert(okay);
+                okay = solver->set("num_original_clauses", getSolverSetup().numOriginalClauses); assert(okay);
 
 		// Check that a version of CaDiCaL is used which has all the unsupported options switched off
 		auto requiredOptionsZero = {"elim", "decompose", "ternary", "vivify", "probe", "transred"};
