### 2022-07-08 7fc6f32
Adding changes missing from https://github.com/domschrei/mallob/commit/c5bf592e621b67e6da602b7fa83ff43bac11dd6c
