+++ src/app/kmeans/kmeans_job.cpp
@@ -25,13 +25,13 @@ void KMeansJob::appl_start() {
 
     countCurrentWorkers = 1;
 
-    LOG(V2_INFO, "                           COMMSIZE: %i myRank: %i myIndex: %i\n",
+    LOG(V3_VERB, "                           COMMSIZE: %i myRank: %i myIndex: %i\n",
         countCurrentWorkers, myRank, myIndex);
-    LOG(V2_INFO, "                           Children: %i\n",
+    LOG(V3_VERB, "                           Children: %i\n",
         this->getJobTree().getNumChildren());
     data = getSerializedDescription(0)->data();
     pointsStart = (float*)(data + getDescription().getMetadataSize() + 3 * sizeof(int));
-    LOG(V2_INFO, "                           myIndex: %i getting ready\n", myIndex);
+    LOG(V3_VERB, "                           myIndex: %i getting ready\n", myIndex);
     baseMsg = JobMessage(getId(),
                          getRevision(),
                          epoch,
@@ -39,7 +39,7 @@ void KMeansJob::appl_start() {
 
     loadInstance();
     clusterMembership.assign(pointsCount, -1);
-    LOG(V2_INFO, "                           myIndex: %i Ready!\n", myIndex);
+    LOG(V3_VERB, "                           myIndex: %i Ready!\n", myIndex);
     loaded = true;
     if (iAmRoot) {
         doInitWork();
@@ -51,8 +51,8 @@ void KMeansJob::initReducer(JobMessage& msg) {
     int lIndex = myIndex * 2 + 1;
     int rIndex = lIndex + 1;
 
-    LOG(V2_INFO, "                           myIndex: %i !tempJobTree.hasLeftChild() %i lIndex in %i\n", myIndex, !tempJobTree.hasLeftChild(), (lIndex < countCurrentWorkers));
-    LOG(V2_INFO, "                           myIndex: %i !tempJobTree.hasRightChild() %i rIndex in %i\n", myIndex, !tempJobTree.hasRightChild(), (rIndex < countCurrentWorkers));
+    LOG(V3_VERB, "                           myIndex: %i !tempJobTree.hasLeftChild() %i lIndex in %i\n", myIndex, !tempJobTree.hasLeftChild(), (lIndex < countCurrentWorkers));
+    LOG(V3_VERB, "                           myIndex: %i !tempJobTree.hasRightChild() %i rIndex in %i\n", myIndex, !tempJobTree.hasRightChild(), (rIndex < countCurrentWorkers));
     work.clear();
     workDone.clear();
     workDone.push_back(myIndex);
@@ -61,10 +61,10 @@ void KMeansJob::initReducer(JobMessage& msg) {
         if (lIndex < countCurrentWorkers) {
             auto grandChilds = KMeansUtils::childIndexesOf(lIndex, countCurrentWorkers);
             work.push_back(lIndex);
-            LOG(V2_INFO, "                           myIndex: %i push in %i\n", myIndex, lIndex);
+            LOG(V3_VERB, "                           myIndex: %i push in %i\n", myIndex, lIndex);
             for (auto child : grandChilds) {
                 work.push_back(child);
-                LOG(V2_INFO, "                           myIndex: %i push in %i\n", myIndex, child);
+                LOG(V3_VERB, "                           myIndex: %i push in %i\n", myIndex, child);
             }
         }
     }
@@ -73,10 +73,10 @@ void KMeansJob::initReducer(JobMessage& msg) {
         if ((rIndex < countCurrentWorkers)) {
             auto grandChilds = KMeansUtils::childIndexesOf(rIndex, countCurrentWorkers);
             work.push_back(rIndex);
-            LOG(V2_INFO, "                           myIndex: %i push in %i\n", myIndex, rIndex);
[…]
