### 2022-07-12 909ea3e
add testscripts
