+++ Out16.5.txt
@@ -1,0 +1,192 @@
+k=5 n=8
+c 
+c Mallob -- a parallel and distributed platform for job scheduling, load balancing, and SAT solving
+c Designed by P. Sanders and D. Schreiber 2018-2022
+c Developed by D. Schreiber 2019-2022
+c 
+c 5.129 3                            loaded
+c 7.854 3                            first iteration finished
+c 34.233 3                            Got Result after iter 44
+s SATISFIABLE
+v 1113063424 1147992552 1147865478 1136055213 1121055608 1134730129 1147401444 1141831117 1113337478 1117475931 1132757103 1141160928 1104158956 1104881282 1099685129 1131599251 1032482280 1016868943 1032488494 1017756600 
+v 1012161320 1017508306 1014458052 1014557370 1013949046 1015099838 1011726802 1016670305 1020400947 1021313438 1014234584 1017328292 1021015477 1020909949 1027240808 1025297900 1017427610 1024736132 1031819635 1028100534 
+v 1025229619 1023444993 1033989136 1114109680 1102273738 1066155424 1084813131 1135832614 1138979292 1110594240 1109433637 1137236304 1144877111 1144469263 1121253649 1066191877 1056606538 1075722638 1035011894 1047990535 
+v 1012907084 1015633008 1018843880 1018934239 1017340854 1015783617 1018000501 1016840852 1024008082 1022139080 1022229452 1020142083 1025193350 1025365038 1024839427 1024738520 1022081854 1064078980 1080758880 1057555991 
+v 1046687802 1127135743 1127805874 1128088179 1104471875 1128488567 1135386596 1135499148 1136099967 1115439512 1139568940 1139830736 1139550611 1136349340 1135741104 1139605644 1139828897 1139131728 1136126112 1134780414 
+v 1139664790 1134474344 1132550638 1133908296 1124722401 1133169127 1120137691 1122112506 1093652386 1017219796 1017753850 1019083749 1009972597 1019858650 1025189534 1022706938 1028132062 1011794664 1018382149 1025487976 
+v 1031184547 1029802290 1017397812 1025194771 1033350380 1106084940 1068158146 1040057474 1076732968 1141784758 1114881170 1083966815 1113236667 1141546968 1150412325 1117138242 1119821885 1119206510 1153061448 1141567557 
+v 1118751972 1123997512 1137963980 1153171501 1117232119 1111757654 1118522311 1137879512 1139910170 1110378158 1110979779 1105872080 1137507182 1037112296 1018801014 1036942138 1018151772 1016120274 1016717157 1017062720 
+v 1011187308 1013553902 1009302412 1090035369 1062600876 1126197416 1092849062 1132746974 1107169661 1126256433 1133636665 1132811192 1133802216 1128058036 1137121900 1134043960 1137464422 1134458496 1137712828 1137457870 
+v 1137663264 1137639764 1137631776 1133189149 1137139542 1137590146 1137185970 1133311694 1124374572 1132215184 1131120588 1131112803 1107617045 1122626012 1120966860 1119377322 1085629369 1038102046 1040288368 1020885409 
+v 1016140478 1016095021 1015872800 1015696037 1019857644 1017963703 1017466236 1023705872 1019774312 1022819760 1020360167 1023549306 1022890464 1023597286 1023550572 1024580860 1041120376 1118929896 1123906854 1138250161 
+v 1152869259 1117056426 1111473910 1118190058 1137418742 1139570558 1110107302 1110742392 1105648210 1137141124 1036913393 1017931303 1036636822 1019703388 1017050380 1016640648 1016988918 1012154729 1014244359 1009880720 
[…]
