+++ Out16.10.txt
@@ -1,0 +1,304 @@
+k=10 n=8
+c 
+c Mallob -- a parallel and distributed platform for job scheduling, load balancing, and SAT solving
+c Designed by P. Sanders and D. Schreiber 2018-2022
+c Developed by D. Schreiber 2019-2022
+c 
+c 5.149 3                            loaded
+c 9.972 3                            first iteration finished
+c 46.510 3                            Got Result after iter 37
+s SATISFIABLE
+v 1113063424 1144571528 1152658100 1119586487 1123571336 1122066810 1155653840 1143432999 1121790474 1125956798 1141218800 1149645218 1112695824 1113425318 1108172912 1140162598 1039300566 1015670132 1039604946 1017904324 
+v 1016699365 1010620258 1014034301 1012879550 1014184920 1017967082 1017628184 1017941976 1008812825 1015030000 1009139168 1015469308 1020176166 1020602920 1014862705 1010243710 1016611504 1025715552 1024793012 1031901357 
+v 1013482029 1020326786 1027548096 1033711936 1031740346 1020992024 1024228188 1033228694 1111839084 1076187040 1049215936 1084026268 1144324121 1116066314 1088571602 1001923472 1015293638 1025873424 1024670555 1031833531 
+v 1014031234 1021405534 1028045096 1033920684 1032048098 1020586279 1023805786 1026393586 1049129277 1022966012 1015241623 1057562153 1144594294 1108988516 1080067048 1116146461 1144366308 1152255560 1119837432 1123758577 
+v 1122573177 1156225170 1143644080 1121934508 1126259772 1141225620 1149898316 1113160412 1113906352 1108502516 1140594056 1039710748 1016490011 1039206838 1019467932 1015436683 1010494138 1014265308 1012938896 1010286074 
+v 1018570654 1016789102 1018583656 1007763294 1015449687 1009583858 1014213292 1020118132 1022523880 1112024961 1125851630 1134436696 1100989022 1105268646 1098803122 1127443836 1072993619 1092388480 1046786568 1019739151 
+v 1020463982 1018841100 1016133254 1008546966 1014181502 1019638866 1016835286 1024168556 1018882129 1022301130 1019780187 1026605181 1026343054 1019192117 1025319628 1027033704 1053322677 1028809317 1029132978 1048389540 
+v 1132786092 1096858818 1068140875 1103323554 1132906228 1140144294 1107865326 1134274180 1134593059 1143412011 1133172601 1135033454 1141530562 1139403834 1141462534 1118653348 1143554920 1141990715 1130437120 1119160300 
+v 1129959580 1143195284 1120305226 1107913023 1078190128 1114705788 1093211368 1151946886 1118833341 1112936542 1120704864 1151195968 1143089476 1114273497 1124045034 1111944468 1156347604 1119816636 1115029134 1121289386 
+v 1141029359 1142391302 1113323773 1114101018 1108584671 1140809142 1040197472 1022271016 1040268284 1015934108 1016711400 1008915636 1015683792 1012709874 1016263468 1011313381 1013052408 1018134239 1019504382 1020690082 
+v 1010891800 1015341256 1018937880 1022020699 1031805569 1013368592 1021401502 1027708532 1022376410 1029796672 1014923177 1021612293 1028472646 1041081808 1021454185 1016447910 1055918472 1022324625 1020489340 1017563086 
+v 1008012130 1013803468 1017369360 1018807000 1028434408 1014843461 1018174847 1028337550 1027149714 1028699506 1011315640 1018276808 1026752072 1088302253 1051777082 1028393631 1062313232 1141934612 1108680724 1079371718 
[…]

[…]
