### 2022-07-12 2655832
get 64 core results
