+++ Out5.txt
@@ -1,0 +1,168 @@
+k=5 n=64
+c 
+c Mallob -- a parallel and distributed platform for job scheduling, load balancing, and SAT solving
+c Designed by P. Sanders and D. Schreiber 2018-2022
+c Developed by D. Schreiber 2019-2022
+c 
+c 8.332 13                            loaded
+c 12.317 13                            first iteration finished
+c 25.670 13                            Got Result after iter 44
+s SATISFIABLE
+v 1113063424 1147992255 1147865189 1136055128 1121055500 1134730068 1147401116 1141830986 1113337432 1117475862 1132757028 1141160815 1104158806 1104881116 1099685024 1131599076 1032482170 1016868827 1032488376 1017756474 
+v 1012161159 1017508180 1014457867 1014557184 1013948867 1015099742 1011726646 1016670193 1020400791 1021319474 1014234403 1017328169 1021015314 1020909790 1027240672 1025297781 1017427485 1024736019 1031819538 1028100385 
+v 1025229500 1023444897 1033988997 1114109568 1102273610 1066155318 1084813014 1135832542 1138980348 1110594962 1109433620 1137236376 1144876874 1144469145 1121253778 1066184833 1056578402 1075722714 1035011972 1047990242 
+v 1012907178 1015633068 1018843950 1018934313 1017340920 1015783672 1018000566 1016840914 1024008122 1022139172 1022229534 1020142159 1025193378 1025365066 1024839457 1024738554 1022081942 1064079006 1080759008 1057556056 
+v 1046687895 1127135680 1127805812 1128088220 1104471894 1128488527 1135386745 1135499307 1136100106 1115439447 1139569168 1139831125 1139550952 1136349655 1135741312 1139606110 1139829270 1139132184 1136126386 1134780580 
+v 1139665166 1134474468 1132550072 1133908380 1124722379 1133169118 1120137571 1122112230 1093651478 1017219796 1017753850 1019083750 1009972597 1019858652 1025189538 1022706941 1028132072 1011794664 1018382150 1025487980 
+v 1031184557 1029802300 1017397814 1025194774 1033350388 1106084908 1068158146 1040057478 1076732954 1141784756 1114881149 1083966814 1113236597 1141547024 1150412314 1117138240 1119821892 1119206520 1153061452 1141567562 
+v 1118751979 1123997589 1137964121 1153171511 1117232128 1111757598 1118522336 1137879524 1139910224 1110378146 1110979740 1105872059 1137507192 1037112328 1018801015 1036942164 1018151772 1016120273 1016717156 1017062721 
+v 1011187308 1013553901 1009302412 1090035309 1062600786 1126197486 1092849132 1132747082 1107169730 1126256503 1133636818 1132811336 1133802480 1128058180 1137122233 1134044204 1137464678 1134458817 1137713054 1137458268 
+v 1137663590 1137640186 1137632037 1133189406 1137139950 1137590578 1137186231 1133311894 1124374568 1132215511 1131120912 1131113134 1107616978 1122625952 1120966764 1119377260 1085629277 1038102020 1040288320 1020885410 
+v 1016140486 1016095032 1015872811 1015696044 1019857637 1017963709 1017466236 1023705866 1019774304 1022819743 1020360160 1023549300 1022890448 1023597280 1023550562 1024580860 1041120390 1118929910 1123906925 1138250156 
+v 1152869313 1117056432 1111473880 1118190069 1137418720 1139570560 1110107297 1110742379 1105648218 1137141125 1036913383 1017931302 1036636814 1019703388 1017050379 1016640648 1016988919 1012154730 1014244358 1009880722 
[…]
