+++ Out20.txt
@@ -1,0 +1,455 @@
+k=20 n=64
+c 
+c Mallob -- a parallel and distributed platform for job scheduling, load balancing, and SAT solving
+c Designed by P. Sanders and D. Schreiber 2018-2022
+c Developed by D. Schreiber 2019-2022
+c 
+c 8.058 13                            loaded
+c 22.639 13                            first iteration finished
+c 36.301 13                            Got Result after iter 22
+s SATISFIABLE
+v 1113063424 1160099685 1120431470 1104098186 1124631842 1129330876 1150008586 1120000636 1120693274 1115868136 1147295432 1045805454 1018737414 1045303038 1022872910 1021060726 1014135804 1021943584 1019411176 990202789 
+v 1002215763 1002029898 1000821775 994663547 1022129449 1016112072 1020549597 1025848202 1028705875 1017064630 992804898 1013252946 1010650837 1001193506 1036467189 1019387941 1027369970 1035044160 1034806022 1010790236 
+v 998405532 990016924 998498464 1037082868 1024024402 1020851628 1064978581 1151951052 1116645112 1087091249 1123426974 1131782446 1149964228 1120360687 1123282802 1032173445 1103490256 1066006493 1039507922 1077700576 
+v 1152781082 1118625471 1091451875 1124500086 1103617736 1160305044 1127567876 1122628266 1129156180 1148457268 1150466936 1121225414 1121966286 1116614999 1148585294 1047502812 1022559250 1047048772 1023772138 1006779014 
+v 1005705831 1004689800 1011147948 1021314612 1014246844 1020222378 1014653257 998288804 1021187608 1013535622 1021898830 1015929646 998796819 1005807434 997317224 1022533848 1027988666 1018393522 1008607870 997926842 
+v 1022737056 1029296805 1039501568 1020120774 1019790564 1027480650 1025956604 1036866238 1021238409 1019433016 1027033049 1034873082 1032467358 1005581527 1007250106 1008484397 1007752966 1037073092 1023981606 1021490167 
+v 1064569642 1151735544 1116543874 1086910185 1123220284 1134305008 1149794190 1119997505 1123878063 1116530090 1160179516 1120385080 1105965894 1124753143 1132762458 1149892802 1119783134 1120465458 1115723608 1146966444 
+v 1045678846 1019090156 1045030272 1022221598 1021353024 1015753000 1021261594 1019113014 990244317 1002838658 1006884390 1000644363 992255754 1021124451 1016233002 1018907298 1025284469 1027821624 1017330150 994084334 
+v 1013055845 1011821554 1002655800 1017142303 1015937332 1017788268 1000240863 1018111250 1003917892 1022508775 1024071280 1019726161 1008877919 1014045634 1019303800 1020210634 1030928439 1016682674 1013200910 1025127184 
+v 1026916007 1031667571 1013573582 1021080201 1027145821 1024797990 1030456388 1016769631 1057221593 1144310073 1108870328 1079905655 1116030267 1094662202 1151491306 1119087971 1114114324 1120499724 1140748556 1151494340 
+v 1118996406 1114417012 1120598904 1150884314 1142478051 1113797995 1123124596 1111354430 1149921698 1113123887 1113280822 1109011620 1140071538 1083890531 1080115190 1080599956 1074418746 1108785134 1146646476 1109941793 
+v 1110398162 1105526802 1136322190 1063895208 1060407089 1062328171 1056319916 1070732466 1028744579 1028937290 1028409872 1030240620 1018489722 1007372740 1011876078 1016329337 1024002888 1029571200 1017434888 1022100508 
[…]

[…]
