### 2022-07-14 f3bddf1
break iteration if root doesnt has to do it alone
