+++ src/app/kmeans/kmeans_job.cpp
@@ -121,6 +121,7 @@ void KMeansJob::sendRootNotification() {
     baseMsg.tag = MSG_BROADCAST_DATA;
     baseMsg.payload = clusterCentersToBroadcast(clusterCenters);
     baseMsg.payload.push_back(countCurrentWorkers);
+    LOG(V3_VERB, "                           myIndex: %i sendRootNotification0\n", myIndex);
     MyMpi::isend(getJobTree().getRootNodeRank(), MSG_SEND_APPLICATION_MESSAGE, baseMsg);
 }
 void KMeansJob::doInitWork() {
@@ -182,7 +183,9 @@ void KMeansJob::appl_communicate() {
                 if (!leftDone) leftDone = (currentIndex == (myIndex * 2 + 1));
                 if (!rightDone) rightDone = (currentIndex == (myIndex * 2 + 2));
                 LOG(V3_VERB, "                           myIndex: %i Start Calc\n", myIndex);
-                calcNearestCenter(metric, cI);
+                if (!skipCurrentIter) {
+                    calcNearestCenter(metric, cI);
+                }
                 LOG(V3_VERB, "                           myIndex: %i End Calc childs\n", myIndex);
 
                 calculatingFinished = true;
@@ -200,36 +203,53 @@ void KMeansJob::appl_communicate() {
         leftDone = false;
         rightDone = false;
         LOG(V3_VERB, "                           myIndex: %i all work Finished2!!!\n", myIndex);
-        auto producer = [&]() {
-            calcCurrentClusterCenters();
+        if (!skipCurrentIter) {
+            auto producer = [&]() {
+                calcCurrentClusterCenters();
 
-            // LOG(V3_VERB, "clusterCenters: \n%s\n", dataToString(localClusterCenters).c_str());
-            return clusterCentersToReduce(localSumMembers, localClusterCenters);
-        };
-        (reducer)->produce(producer);
-        (reducer)->advance();
+                // LOG(V3_VERB, "clusterCenters: \n%s\n", dataToString(localClusterCenters).c_str());
+                return clusterCentersToReduce(localSumMembers, localClusterCenters);
+            };
+            (reducer)->produce(producer);
+            (reducer)->advance();
+        }
         calculatingFinished = false;
         LOG(V3_VERB, "                           myIndex: %i all work Finished END!!!\n", myIndex);
     };
     if (hasReducer) {
-        if (iAmRoot && (reducer)->hasResult()) {
-            LOG(V3_VERB, "                           myIndex: %i received Result from Transform\n", myIndex);
-            clusterCenters = broadcastToClusterCenters((reducer)->extractResult(), true);
-            clusterMembership.assign(pointsCount, -1);
+        if (!skipCurrentIter) {
+            if (iAmRoot && (reducer)->hasResult()) {
+                LOG(V3_VERB, "                           myIndex: %i received Result from Transform\n", myIndex);
+                clusterCenters = broadcastToClusterCenters((reducer)->extractResult(), true);
+                clusterMembership.assign(pointsCount, -1);
 
+                baseMsg.tag = MSG_BROADCAST_DATA;
+                baseMsg.payload = clusterCentersToBroadcast(clusterCenters);
+                baseMsg.payload.push_back(countCurrentWorkers);
+                LOG(V3_VERB, "                           myIndex: %i sendRootNotification1\n", myIndex);
+                MyMpi::isend(getJobTree().getRootNodeRank(), MSG_SEND_APPLICATION_MESSAGE, baseMsg);
+                // only root...?
+
+                // continue broadcasting
+
[…]
