### 2022-07-14 9e42ecf
Add STXXL dependency and simple ext. priority queue wrapper
