+++ lib/fetch_stxxl.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+git clone https://github.com/stxxl/stxxl.git
+cd stxxl
+git checkout b9e44f0ecba7d7111fbb33f3330c3e53f2b75236
