### 2022-07-14 2fc2c19
Add size() to external priority queue
