+++ instances/proofs/README.md
@@ -1,0 +1,1 @@
+These example LRAT proof files are taken from https://github.com/marijnheule/drat-trim
