+++ CMakeLists.txt
@@ -167,3 +167,4 @@ new_test(concurrent_malloc)
 new_test(distributed_clause_filter)
 new_test(hashing)
 new_test(ext_priority_queue)
+new_test(reverse_lrat_parser)
