### 2022-07-15 ae9166b
Add commands to fetch_stxxl.sh to generate config.h
