### 2022-07-15 1b2fb8a
Continue implementation of distributed proof combination (ProofInstance untested)
