+++ src/app/sat/proof/proof_instance.hpp
@@ -1,0 +1,195 @@
+
+#pragma once
+
+#include <algorithm>
+
+#include "app/sat/util/reverse_lrat_parser.hpp"
+#include "util/external_priority_queue.hpp"
+#include "util/sys/thread_pool.hpp"
+
+class ProofInstance {
+
+private:
+    const int _instance_id;
+    const int _num_instances;
+    bool _winning_instance;
+
+    ReverseLratParser _parser;
+    ReverseLratParser::LratLine _current_line;
+    ExternalPriorityQueue<LratClauseId> _frontier;
+    ExternalPriorityQueue<LratClauseId> _backlog;
+
+    std::vector<LratClauseId> _local_epoch_starts;
+    std::vector<LratClauseId> _local_epoch_offsets;
+    std::vector<LratClauseId> _global_epoch_starts;
+    int _current_epoch;
+
+    std::ofstream _output;
+
+    std::future<void> _work_future;
+    bool _work_done = true;
+    std::vector<LratClauseId> _outgoing_clause_ids;
+
+public:
+    ProofInstance(int instanceId, int numInstances, 
+        const std::string& proofFilename, int finalEpoch, 
+        int winningInstance, const std::vector<LratClauseId>& localEpochStarts, 
+        const std::vector<LratClauseId>& localEpochOffsets,
+        const std::string& outputFilename) :
+            _instance_id(instanceId), _num_instances(numInstances),
+            _winning_instance(winningInstance == instanceId),
+            _parser(proofFilename), _local_epoch_starts(localEpochStarts), 
+            _local_epoch_offsets(localEpochOffsets), _current_epoch(finalEpoch),
+            _output(outputFilename) {
+        
+        _global_epoch_starts.resize(_local_epoch_starts.size());
+        for (size_t i = 0; i < _global_epoch_starts.size(); ++i) {
+            _global_epoch_starts[i] = _local_epoch_starts[i] + _local_epoch_offsets[i];
+        }
+    }
+
+    void advance(const std::shared_ptr<std::vector<LratClauseId>>& incomingClauseIds) {
+        _work_done = false;
+        _work_future = ProcessWideThreadPool::get().addTask([this, incomingClauseIds]() {
+            handleIncomingClauseIds(*incomingClauseIds);
+            readEpoch();
+            prepareNextOutgoingClauseIds();
+            _work_done = true;
+        });
+    }
+
+    bool ready() const {return _work_done;}
+
+    std::vector<LratClauseId>&& extractNextOutgoingClauseIds() {
+        assert(ready());
+        _work_future.get();
+
+        // Proceed with the next (earlier) epoch
+        assert(_current_epoch > 0);
+        _current_epoch--;
+
+        return std::move(_outgoing_clause_ids);
+    }
+
+private:
+
+    void handleIncomingClauseIds(const std::vector<LratClauseId>& clauseIds) {
+        
+        // Import self-produced clauses
+        for (auto id : clauseIds) {
+            if (isSelfProducedClause(id)) _frontier.push(id);
+        }
+    }
+
+    void readEpoch() {
+        if (!_current_line.valid() && _parser.hasNext()) 
+            _current_line = _parser.next();
+        
+        while (_current_line.valid()) {
+
+            alignSelfProducedClauseIds(_current_line);
+            auto id = _current_line.id;
+
+            int epoch = getClauseEpoch(id);
+            if (epoch != _current_epoch) {
+                // check if it is from a future epoch (which would be an error)
+                assert(epoch < _current_epoch);
+                // stop reading because a former epoch has been reached
+                break; 
+            }
+
+            if (_frontier.top() == id || 
+                    (_current_line.literals.empty() && _winning_instance)) {
[…]
