### 2022-07-15 67f5537
More reverse_lrat_parser.hpp into proof directory
