### 2022-07-15 66bf7ba
better test.bash, cleanup files, bugfix in break
