### 2022-07-19 1f9e657
Implement alignment of clause IDs and gathering of epoch-wise ID ranges
