+++ src/app/sat/sharing/sharing_manager.cpp
@@ -6,6 +6,8 @@
  */
 
 #include <signal.h>
+#include <algorithm>
+#include <fstream>
 
 #include "util/assert.hpp"
 
@@ -42,11 +44,16 @@ SharingManager::SharingManager(
 
 	auto callback = getCallback();
 	
-    for (size_t i = 0; i < _solvers.size(); i++) {
+	unsigned long zerothEpochIdStart = solvers.empty() ? 0 : 
+		solvers[0]->getSolverSetup().numOriginalClauses+1;
+	for (size_t i = 0; i < _solvers.size(); i++) {
 		_solvers[i]->setExtLearnedClauseCallback(callback);
 		_solver_revisions.push_back(_solvers[i]->getSolverSetup().solverRevision);
 		_solver_stats.push_back(&_solvers[i]->getSolverStatsRef());
+		_id_offsets_per_solver[i].push_back(0);
+		_min_epoch_ids_per_solver[i].push_back(zerothEpochIdStart);
 	}
+	_global_epoch_ids.push_back(zerothEpochIdStart);
 }
 
 void SharingManager::onProduceClause(int solverId, int solverRevision, const Clause& clause, int condVarOrZero) {
@@ -59,6 +66,12 @@ void SharingManager::onProduceClause(int solverId, int solverRevision, const Cla
 			LOGGER(_logger, V3_VERB, "Simulating a crash!\n");
 			raise(SIGSEGV); // causes crash
 		}
+	}
+
+	if (MALLOB_CLAUSE_METADATA_SIZE == 2) {
+		unsigned long clauseId = metadata::readUnsignedLong(clause.begin);
+		_last_exported_clause_id[solverId]->store(clauseId, std::memory_order_relaxed);
+		assert(getProducingLocalSolverIndex(clauseId) == solverId);
 	}
 
 	auto clauseBegin = clause.begin;
@@ -136,7 +149,12 @@ void SharingManager::onProduceClause(int solverId, int solverRevision, const Cla
 int SharingManager::prepareSharing(int* begin, int totalLiteralLimit) {
 
 	int numExportedClauses = 0;
-	auto buffer = _cdb.exportBuffer(totalLiteralLimit, numExportedClauses);
+	auto buffer = _cdb.exportBuffer(totalLiteralLimit, numExportedClauses, 
+			AdaptiveClauseDatabase::ANY, /*sortClauses=*/true, [&](int* data) {
+
+		// Shift clause ID from a local solver according to the solver's offset
+		if (MALLOB_CLAUSE_METADATA_SIZE == 2) alignClauseId(data);
+	});
 	//assert(buffer.size() <= maxSize);
 	memcpy(begin, buffer.data(), buffer.size()*sizeof(int));
 
@@ -152,6 +170,11 @@ void SharingManager::returnClauses(int* begin, int buflen) {
 	auto reader = _cdb.getBufferReader(begin, buflen);
 	auto c = reader.getNextIncomingClause();
 	while (c.begin != nullptr) {
+
+		// Returned clauses would be aligned *again* when re-exported.
+		// => subtract the offsets again here ...
+		if (MALLOB_CLAUSE_METADATA_SIZE == 2) unalignClauseId(c.begin);
+
 		bool success = _cdb.addClause(c);
 		if (success) _hist_returned_to_db.increment(c.size);
 		c = reader.getNextIncomingClause();
@@ -165,9 +188,28 @@ int SharingManager::filterSharing(int* begin, int buflen, int* filterOut) {
 	constexpr auto bitsPerElem = 8*sizeof(int);
 	int shift = bitsPerElem;
 	auto clause = reader.getNextIncomingClause();
-	int filterPos = -1;
[…]
