+++ src/app/sat/data/clause_metadata_def.hpp
@@ -1,6 +1,8 @@
 
 #pragma once
 
+#include <cstring>
+
 #ifndef MALLOB_CLAUSE_METADATA_SIZE
 #define MALLOB_CLAUSE_METADATA_SIZE 0
 #endif
@@ -8,3 +10,14 @@
 static_assert(MALLOB_CLAUSE_METADATA_SIZE == 0 || MALLOB_CLAUSE_METADATA_SIZE >= 2, 
     "MALLOB_CLAUSE_METADATA_SIZE must either be zero or at least two "
     "since clauses of length two including metadata are not yet supported.");
+
+namespace metadata {
+	unsigned long readUnsignedLong(const int* data) {
+		unsigned long ul;
+		memcpy(&ul, data, sizeof(unsigned long));
+		return ul;
+	}
+	void writeUnsignedLong(unsigned long ul, int* data) {
+		memcpy(data, &ul, sizeof(unsigned long));
+	}
+}
