### 2022-07-19 91e414f
Write clause epochs file upon final output of statistics in SAT engine
