### 2022-07-19 5049d30
#7 Document OpenMPI requirement for test suite
