### 2022-07-19 f8e0e0f
Add JOSS status badge
