### 2022-07-20 0e1955d
Add option for turning off memory panic
