### 2022-07-20 331387d
Add option for turning off memory panic
