### 2022-07-20 #4591 @danielskatz
> Regarding the invalid DOI, this is a figshare reference to our software reproducibility artifact which got us invited to the Euro-Par special issue. The DOI is not active yet, but it should become active together with the Euro-Par '22 proceedings.

ok, thanks - let's leave it as is then
