### 2022-07-22 bb94120
Fix clause comparison bug with clause IDs; safety checks and forced termination if child terminates in case of certified UNSAT
