+++ src/app/sat/sharing/import_buffer.hpp
@@ -46,6 +46,18 @@ public:
     }
 
     void add(const Mallob::Clause& c) {
+        if (MALLOB_CLAUSE_METADATA_SIZE == 2) {
+            // Perform various safety checks
+            assert(c.size >= 3);
+            unsigned long id; memcpy(&id, c.begin, sizeof(unsigned long));
+            assert(id < std::numeric_limits<unsigned long>::max()/2
+                    || log_return_false("Clause ID \"%lu\" found, which could be an error\n", id));
+            for (size_t i = MALLOB_CLAUSE_METADATA_SIZE; i < c.size; i++)
+                assert(std::abs(c.begin[i]) < 10000000 
+                    || log_return_false("Literal \"%i\" found, error thrown for safety - "
+                    "delete this assertion if your formula has >=10'000'000 variables\n", 
+                    c.begin[i]));
+        }
         bool success = _cdb.addClause(c);
         if (!success) _stats.receivedClausesDropped++;
     }
