### 2022-07-22 #4591 @massimotorquati
Yes, I completed the missing points. I have no objections; the work is OK with me.
