+++ paper/paper.md
@@ -25,21 +25,21 @@ bibliography: paper.bib
 
 # Summary
 
-Propositional satisfiability (SAT) is the problem of finding a variable assignment for a given propositional formula (i.e., a composition of Boolean variables using logical operators NOT, AND, OR) such that the formula evaluates to **true**, or reporting that no such assignment exists. The platform **Mallob** (**Mal**leable **Lo**ad **B**alancer, or **Ma**ssively P**a**ra**ll**el **Lo**gic **B**ackend) allows to process SAT jobs in a (massively) parallel and distributed system on demand. Mallob's flexible, fair, and decentralized approach to online job scheduling results in scheduling latencies in the range of milliseconds, near-optimal system utilization, and high resource efficiency.
+Propositional satisfiability (SAT) is the problem of finding a variable assignment for a given propositional formula (i.e., a composition of Boolean variables using logical operators NOT, AND, OR) such that the formula evaluates to **true**, or reporting that no such assignment exists. The platform **Mallob** (**Mal**leable **Lo**ad **B**alancer, or **Ma**ssively P**a**ra**ll**el **Lo**gic **B**ackend) enables processing SAT jobs in a (massively) parallel and distributed system on demand. Mallob's flexible, fair, and decentralized approach to online job scheduling results in scheduling latencies in the range of milliseconds, near-optimal system utilization, and high resource efficiency.
 
 # Statement of need
 
 Despite SAT being a notoriously difficult problem [@cook1971complexity], practically efficient SAT solving approaches have empowered a wide range of real-world applications for SAT such as software verification [@buning2019using], circuit design [@goldberg2001using], cryptography [@massacci2000logical], automated planning [@schreiber2021lilotane], and theorem proving [@heule2016solving].
 With respect to modern computing paradigms such as cloud computing and high-performance computing (HPC), the limited scalability of established parallel and distributed SAT solvers has become a pressing issue [@hamadi2012seven].
-Moreover, since processing times of SAT jobs are unknown in advance, conventional HPC scheduling approaches applied to such jobs can lead to prohibitively large scheduling latencies and to suboptimal utilization of computational resources. Instead, we suggest to make use of so-called _malleable scheduling_. A parallel computing task is called malleable if the amount of computational resources allotted (i.e., the number of cores or machines) can be adjusted during its execution [@feitelson1997job]. Malleability is a powerful paradigm in the field of job scheduling and load balancing as it allows to schedule incoming jobs rapidly and to continuously rebalance the present tasks according to their priority and other properties. 
+Moreover, since processing times of SAT jobs are unknown in advance, conventional HPC scheduling approaches applied to such jobs can lead to prohibitively large scheduling latencies and to suboptimal utilization of computational resources. Instead, we suggest making use of so-called _malleable scheduling_. A parallel computing task is called malleable if the amount of computational resources allotted (i.e., the number of cores or machines) can be adjusted during its execution [@feitelson1997job]. Malleability is a powerful paradigm in the field of job scheduling and load balancing as it allows scheduling incoming jobs rapidly and continuously rebalancing the present tasks according to their priority and other properties. 
 
-We believe that a cloud service which combines a scalable SAT solving engine with malleable job scheduling can significantly improve many SAT-related academic and industrial workflows in terms of productivity and (resource-)efficiency. Moreover, our decentralized scheduling approach is applicable to further applications beyond SAT where processing times are unknown in advance and a modest amount of data is transferred between the workers.
[…]
