### 2022-07-22 #4591 @danielskatz
@domschrei - I'm suggesting some minor changes in https://github.com/domschrei/mallob/pull/8. (note that this includes removing the funding agency logo at the end, which would be appropriate on slides or a poster, but not in a paper)  Please merge this, or let me know what you disagree with, and we can move forward on acceptance.
