### 2022-07-23 #4591 @domschrei
I created a tagged release [v1.1.0](https://github.com/domschrei/mallob/releases/tag/v1.1.0) and uploaded that version on Zenodo. The DOI is: https://doi.org/10.5281/zenodo.6890240
