### 2022-07-23 7f7c688
Remove duplicate bibliography entry
