+++ paper/paper.bib
@@ -123,17 +123,6 @@
 	publisher={Citeseer}
 }
 
-@article{froleyks2021sat,
-	title={{SAT} competition 2020},
-	author={Froleyks, Nils and Heule, Marijn and Iser, Markus and J{\"a}rvisalo, Matti and Suda, Martin},
-	journal={Artificial Intelligence},
-	volume={301},
-	pages={103572},
-	year={2021},
-	publisher={Elsevier},
-	doi={10.1016/j.artint.2021.103572}
-}
-
 @inproceedings{goldberg2001using,
   title={Using SAT for combinational equivalence checking},
   author={Goldberg, Evguenii I and Prasad, Mukul R and Brayton, Robert K},
