### 2022-07-23 #4591 @danielskatz
Current status:
* paper is proofread, and content is ready to publish
* missing DOI is ok, it's a DOI that will be associate with a to-be-published paper
* As a Euro-Par associated paper, we need to fix the left column to add appropriate language and logo before final publishing

We're now on hold until the last item is figured out.
