+++ Testing/test.sh
@@ -4,7 +4,7 @@ folder="latestResults" #latestResults
 instanceName="covtypeShuffle" #   mnist784      benign_traffic      covtypeShuffle
 instanceFirstLine="54 55 550000" #7 7 209    115 115 52150      54 55 581012
 kList=(20 30 40 50) # 60 70 80
-npList=(160 155 150 145 140 135 130 128 127 125 120 115 110 105 100 95 90 85 80 75 70 65 64 63 60 55 50 45 40 35 32 31 30 25 20 16 15 10 8 7 5 4 3 2 1) 
+npList=(65 64 63 60 55 50 45 40 35 32 31 30 25 20 16 15 10 8 7 5 4 3 2 1) # 160 155 150 145 140 135 130 128 127 125 120 115 110 105 100 95 90 85 80 75 70 
 countPasses=3
 #k=20
 for k in ${kList[@]}; do
@@ -12,7 +12,7 @@ for k in ${kList[@]}; do
 
     echo "$k $instanceFirstLine" > ./instances/${instanceName}${k}.csv
     cat ./instances/${instanceName}K.csv >> ./instances/${instanceName}${k}.csv
-    > ./Testing/${folder}/times-${pcName}-${k}.txt
+#    > ./Testing/${folder}/times-${pcName}-${k}.txt
     for n in ${npList[@]}; do #((n = 158 ; n < 161 ; ++n));
         # v=3 
         #PATH=build/:$PATH RDMAV_FORKSAVE=1 mpirun -np ${n} -oversubscribe build/mallob -mono-application=KMEANS -mono=./instances/${instanceName}${k}.csv -v=3
