+++ Testing/test.sh
@@ -1,11 +1,11 @@
 #!/bin/bash
-pcName="i10pc135"
-folder="dTest" #latestResults
+pcName="i10pc138"
+folder="dTest256core" #latestResults
 instanceName="covtypeShuffle" #   mnist784      benign_traffic      covtypeShuffle
 instanceFirstLine="54 55" #7 7 209    115 115 52150      54 55 581012
 kList=(30) # 60 70 80
-npList=(160 155 150 145 140 135 130 128 127 125 120 115 110 105 100 95 90 85 80 75 70 65 64 63 60 55 50 45 40 35 32 31 30 25 20 16 15 10 8 7 5 4 3 2 1) #   65 64 63 60 55 50 45 40 35 32 31 30 25 20 16 15 10 8 7 5 4 3 2 1
-dList=(10000 100000 250000 500000)
+npList=(256 255 250 245 240 235 230 225 220 215 210 205 200 195 190 185 180 175 170 165 160 155 150 145 140 135 130 128 127 125 120 115 110 105 100 95 90 85 80 75 70 65 64 63 60 55 50 45 40 35 32 31 30 25 20 16 15 10 8 7 5 4 3 2 1) #   65 64 63 60 55 50 45 40 35 32 31 30 25 20 16 15 10 8 7 5 4 3 2 1
+dList=(10000 50000 100000 300000 500000)
 countPasses=2
 
 for k in ${kList[@]}; do
