### 2022-07-26 e87f82c
change speedup and efficiency calc
