+++ src/app/sat/execution/engine.cpp
@@ -21,6 +21,7 @@
 #include <csignal>
 #include <unistd.h>
 #include <sched.h>
+#include <filesystem>
 #include "util/assert.hpp"
 
 using namespace SolvingStates;
@@ -95,6 +96,12 @@ SatEngine::SatEngine(const Parameters& params, const SatProcessConfig& config, L
 	setup.certifiedUnsat = params.certifiedUnsat();
         setup.maxNumSolvers = config.mpisize * params.numThreadsPerProcess();
         setup.numOriginalClauses = numClauses;
+        //make a directory for putting proofs in
+        std::filesystem::path base_dir(params.logDirectory());
+        std::filesystem::path proof_dir("proof");
+        std::filesystem::path full_path = base_dir / proof_dir;
+        create_directory(full_path);
+        setup.proofDir = full_path.string();
 
 	// Instantiate solvers according to the global solver IDs and diversification indices
 	int cyclePos = begunCyclePos;
