+++ src/app/sat/execution/solver_setup.hpp
@@ -40,4 +40,5 @@ struct SolverSetup {
 	bool skipClauseSharingDiagonally;
 	bool certifiedUnsat;
         int maxNumSolvers;
+        std::string proofDir;
 };
