### 2022-08-04 4638287
kill hung mallob processes
