### 2022-08-04 cdd75db
Generalize job streamer to multiple input files per job
