### 2022-08-05 #4591 @arfon
@danielskatz – we need a PR to the `paper.md` for this submission that looks like this: https://github.com/yogi-tud/space_gpu/pull/5/files

Then you need to do `@editorialbot accept`

I don't think I know what the Euro-Par DOI is though for this submission?
