### 2022-08-05 #4591 @domschrei
Maybe I can help. There's two DOIs associated with this work:

* DOI of the Figshare repository of our Euro-Par artifact: https://doi.org/10.6084/m9.figshare.20000642 (not active yet)
* DOI of the Euro-Par publication itself as given by Springer: https://doi.org/10.1007/978-3-031-12597-3_8 (active)
