+++ Testing/test.sh
@@ -4,8 +4,8 @@ folder="nTest256core100d" #latestResults
 instanceName="benign_trafficShuffle" #   mnist784      benign_trafficShuffle      covtypeShuffle
 instanceFirstLine="115 115" #              7 7 209       115 115 52150              54 55 581012
 kList=(50) # 60 70 80
-wList=(140 130 128 127 125 120 110 100 90 80 75 70 65 64 63 60 55 50 45 40 35 32 31 30 25 20 16 15 10 8 7 5 4 3 2 1) #   65 64 63 60 55 50 45 40 35 32 31 30 25 20 16 15 10 8 7 5 4 3 2 1
-nList=(10000 30000 50000)
+wList=(40 35 32 31 30 25 20 16 15 10 8 7 5 4 3 2 1) #   65 64 63 60 55 50 45 40 35 32 31 30 25 20 16 15 10 8 7 5 4 3 2 1
+nList=(50000)
 countPasses=2
 
 for k in ${kList[@]}; do
