### 2022-08-06 #4591 @domschrei
And regarding the PR @arfon referenced, the DOI of the publication itself seems to be the correct one to insert. Should I just go ahead and add this line to `paper.md`?
