+++ Testing/test.sh
@@ -3,9 +3,9 @@ pcName="i10pc138"
 folder="nTest256core50dNochmal" #latestResults
 instanceName="covtypeShuffle" #   mnist784      benign_trafficShuffle      covtypeShuffle
 instanceFirstLine="54 55" #              7 7 209       115 115 52150              54 55 581012
-kList=(10 30 50 70 100) # 60 70 80
+kList=(100) # 60 70 80
 wList=(160 155 150 145 140 130 128 127 125 120 110 100 90 80 75 70 65 64 63 60 55 50 45 40 35 32 31 30 25 20 16 15 10 8 7 5 4 3 2 1) #   65 64 63 60 55 50 45 40 35 32 31 30 25 20 16 15 10 8 7 5 4 3 2 1
-nList=(100000 300000 500000)
+nList=(300000 500000)
 countPasses=1
 
 ./Testing/killHung.sh ${folder} &
