+++ src/app/kmeans/kmeans_job.hpp
@@ -29,6 +29,8 @@ class KMeansJob : public Job {
     int iterationsDone = 0;
     int epoch = 0;
     int countReady = 0;
+    int maxDemand = -1;
+    bool maxDemandCalculated = false;
     std::vector<Point> kMeansData;
     uint8_t* data;
     const float* pointsStart;
@@ -78,7 +80,7 @@ class KMeansJob : public Job {
             LOG(V3_VERB, "                           AllCollected: Error\n");
         }
         auto transformed = clusterCentersToBroadcast(clusterCenters);
-        transformed.push_back(this->getVolume()); // will be countCurrentWorkers
+        transformed.push_back(this->getVolume());  // will be countCurrentWorkers
         LOG(V3_VERB, "                           COMMSIZE: %i myIndex: %i \n",
             this->getVolume(), myIndex);
         LOG(V3_VERB, "                           Children: %i\n",
@@ -131,6 +133,38 @@ class KMeansJob : public Job {
     void appl_dumpStats() override;
     bool appl_isDestructible() override { return true; }
     void appl_memoryPanic() override;
+    int getDemand() {
+        if (!loaded) {
+            return Job::getDemand();
+        } else {
+            if (maxDemandCalculated) {
+                return maxDemand;
+            } else {
+                double problemSize = countClusters * dimension * pointsCount;
+                maxDemand = -330.7589 + 20.2719 * log(problemSize);  // evaluated with tests
+                maxDemandCalculated = true;
+
+                if (maxDemand < 12) {
+                    maxDemand = 7;
+                } else {
+                    uint32_t positiveDemand = maxDemand;
+                    int lowerBound = 1;
+                    while (positiveDemand >>= 1) {
+                        lowerBound <<= 1;
+                    }
+
+                    int middle = lowerBound * 1.5;
+                    if (maxDemand > middle) {
+                        maxDemand = lowerBound * 2 - 1;
+                    } else {
+                        maxDemand = lowerBound - 1;
+                    }
+                }
+                maxDemand = std::min(maxDemand, Job::getGlobalNumWorkers());
+                return maxDemand;
+            }
+        }
+    }
 
     void loadInstance();
     void doInitWork();
