### 2022-08-09 8fe71e2
Add reference to Euro-Par '22 paper
