### 2022-08-09 166da0b
Yet untested code for distributed proof assembly (without merging of files)
