+++ src/app/sat/proof/proof_assembler.hpp
@@ -1,0 +1,163 @@
+
+#include "proof_instance.hpp"
+
+class ProofAssembler {
+
+private:
+    const Parameters& _params;
+    int _job_id;
+    int _num_workers;
+    int _threads_per_worker;
+    int _this_worker_index;
+    int _final_epoch;
+    int _winning_instance;
+    std::list<ProofInstance> _proof_instances;
+
+    int _current_epoch;
+
+public:
+    ProofAssembler(const Parameters& params, int jobId, int numWorkers, int threadsPerWorker, 
+        int thisWorkerIndex, int finalEpoch, int winningInstance) :
+            _params(params), _job_id(jobId), _num_workers(numWorkers), _threads_per_worker(threadsPerWorker), 
+            _this_worker_index(thisWorkerIndex), _final_epoch(finalEpoch), _winning_instance(winningInstance) {
+
+        createInstancesViaClauseEpochs(params.logDirectory() + "/proof#" + std::to_string(jobId) 
+            + "/clauseepochs." + std::to_string(thisWorkerIndex));
+        _current_epoch = _final_epoch;
+        beginProofAssembly();
+    }
+
+    int getEpoch() const {
+        return _current_epoch;
+    }
+
+    bool finished() const {
+        for (const auto& inst : _proof_instances) {
+            if (!inst.finished()) return false;
+        }
+        return true;
+    }
+
+    bool canEmitClauseIds() const {
+        bool someReady = false;
+        for (const auto& inst : _proof_instances) {
+            if (!inst.finished()) {
+                if (inst.ready()) someReady = true;
+                else return false;
+            }
+        }
+        return someReady;
+    }
+
+    std::vector<LratClauseId> emitClauseIds() {
+        std::list<std::vector<LratClauseId>> instancesClauseLists;
+        std::list<std::pair<LratClauseId*, size_t>> instancesIdArrays;
+        for (auto& inst : _proof_instances) {
+            instancesClauseLists.push_back(inst.extractNextOutgoingClauseIds());
+            instancesIdArrays.emplace_back(instancesClauseLists.back().data(), instancesClauseLists.back().size());
+        }
+        return mergeClauseIdVectors(instancesIdArrays);
+    }
+
+    void importClauseIds(const LratClauseId* clauseIdsData, size_t clauseIdsSize) {
+        _current_epoch--;
+        for (auto& inst : _proof_instances) {
+            inst.advance(clauseIdsData, clauseIdsSize);
+        }
+    }
+
+    std::vector<LratClauseId> mergeClauseIdVectors(const std::list<std::pair<LratClauseId*, size_t>>& idArrays) {
+        
+        std::vector<LratClauseId> output;
+        
+        std::vector<LratClauseId*> pointers;
+        std::vector<size_t> numRemaining;
+        for (auto& [pointer, size] : idArrays) {
+            pointers.push_back(pointer);
+            numRemaining.push_back(size);
+        }
+
+        while (true) {
+            bool anyLeft = false;
+            size_t nextVecPos;
+            LratClauseId nextClauseId;
+
+            for (size_t i = 0; i < pointers.size(); i++) {
+                auto& pointer = pointers[i];
+                auto& remaining = numRemaining[i];
+                if (remaining == 0) continue;
+                auto clauseId = *pointer;
+                if (!anyLeft || clauseId < nextClauseId) {
+                    anyLeft = true;
+                    nextVecPos = i;
+                    nextClauseId = clauseId;
+                }
+            }
+            if (!anyLeft) break;
+
+            output.push_back(nextClauseId);
+            numRemaining[nextVecPos]--;
+        }
+
+        return output;
+    }
+
+private:
+    void createInstancesViaClauseEpochs(const std::string& filename) {
+
[…]
