+++ src/app/sat/execution/engine.cpp
@@ -258,6 +258,7 @@ int SatEngine::solveLoop() {
 			if (result.result > 0 && result.revision == _revision) {
 				done = true;
 				_result = std::move(result);
+				_result.localWinningInstanceId = i;
 				break;
 			}
 		}
