+++ src/data/job_result.hpp
@@ -15,6 +15,7 @@ struct JobResult : public Serializable {
     int revision;
     int result;
     enum EncodedType {INT, FLOAT} encodedType = INT;
+    int localWinningInstanceId;
 
 private:
     std::vector<int> solution;

[…]
